Start using CGFloat
Fixes a bunch of compiler warnings.  Since CGFloat does not exist on
Tiger and earlier, we have added a typedef likewise as was done for
NSInteger.

+++ gdk/quartz/gdkcolor-quartz.c
@@ -199,10 +199,10 @@ gdk_colormap_get_screen (GdkColormap *cmap)
 void
 _gdk_quartz_colormap_get_rgba_from_pixel (GdkColormap *colormap,
 					  guint32      pixel,
-					  float       *red,
-					  float       *green,
-					  float       *blue,
-					  float       *alpha)
+					  CGFloat      *red,
+					  CGFloat      *green,
+					  CGFloat      *blue,
+					  CGFloat      *alpha)
 {
   *red   = (pixel >> 16 & 0xff) / 255.0;
   *green = (pixel >> 8  & 0xff) / 255.0;
